pyproject.toml
[project]
name = "discovery-dojo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "openai>=1.97.0",
    "pocketflow>=0.0.2",
    "pre-commit>=4.2.0",
    "ruff>=0.12.3",
    "tavily-python>=0.7.10",
]

[tool.ruff.lint.per-file-ignores]
"__init__.pyi" = ["F401"]
